Add nil safety and slice/map safety guidelines; refactor error handling and improve context usage
- Introduced comprehensive documentation on nil safety in Go, covering nil pointer receivers, function values, error comparisons, and generic code.
- Added guidelines for slice and map safety, including range loop variable capture, subslice retention, and defensive copying.
- Refactored error handling in the API to provide more context in error messages.
- Improved context usage in various handlers to ensure proper type assertions and error handling.
- Updated SQL migration scripts for comments table to use BIGINT for IDs.
- Enhanced user and post storage methods to return more informative error messages.
- Updated skills-lock.json to include new golang-safety skill.

+++ internal/db/seed.go
@@ -247,7 +247,7 @@ func generatePosts(num int, users []*store.User) []*store.Post {
 
 	for i := range num {
 		posts[i] = &store.Post{
-			UserId:  int64(users[i%len(users)].ID),
+			UserId:  users[i%len(users)].ID,
 			Title:   postTitles[i],
 			Content: postContents[i],
 			Tags:    postTags[i],
@@ -262,7 +262,7 @@ func generateComments(num int, posts []*store.Post, users []*store.User) []*stor
 	for i := range num {
 		comments = append(comments, &store.Comment{
 			PostID:  posts[i%len(posts)].ID,
-			UserID:  int64(users[(i+3)%len(users)].ID),
+			UserID:  users[(i+3)%len(users)].ID,
 			Content: commentTexts[i%len(commentTexts)],
 		})
 	}
